+++ R/ggeffect.R
@@ -90,7 +90,7 @@ ggeffect_helper <- function(model, terms, ci.lvl, x.as.factor, ...) {
   # build data frame, with raw values
   # predicted response and lower/upper ci
 
-  if (inherits(model, c("polr", "clm", "multinom"))) {
+  if (inherits(model, c("polr", "clm", "clm2", "multinom"))) {
 
     # for categorical outcomes, we need to gather the data
     # from effects to get a single data frame